+++ src/Video/Exception/UnexpectedValueException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Soluble\MediaTools\Video\Exception;
+
+use Soluble\MediaTools\Common\Exception\UnexpectedValueException as UVE;
+
+class UnexpectedValueException extends UVE
+{
+}