+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Added `ext-pcre` to required deps
+- Minor, added `UnexpectedValueException` to handle case when ffmpeg cli command cannot be generated.
 
 ## 0.8.0 (2018-08-27)